BUGFIX: Correct benchmark_fncs typing

+++ glompo/benchmark_fncs/griewank.py
@@ -8,7 +8,7 @@ from ._base import BaseTestCase
 class Griewank(BaseTestCase):
     """ When called returns evaluations of the Griewank function. """
 
-    def __init__(self, dims: int = 2, delay: int = 0):
+    def __init__(self, dims: int = 2, delay: float = 0):
         """
         Implementation of the Griewank optimization test function.
         Recommended bounds: [-600, 600] * dims
@@ -19,7 +19,7 @@ class Griewank(BaseTestCase):
         ----------
         dims: int = 2
             Number of dimensions of the problem
-        delay: int = 0
+        delay: float = 0
             Delay in seconds after the function is called before results are returned.
             Useful to simulate harder problems.
         """
